Load nsrt file in RuleTest to avoid duplicate fixture
